requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
PyMuPDF==1.24.9
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
transformers==4.44.2
torch>=2.0.0
numpy<2.0
